requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
rasterio>=1.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
joblib>=1.1.0
streamlit>=1.0.0
imageio>=2.0.0
pillow>=9.0.0
scipy>=1.7.0